Remove post 23
Sorry, this was causing more stress than a single published piece of art
should.

It's obviously still here in this git repository and I don't want to
pretend I didn't draw it. I just don't want it to be something I..
present. It's a personal idea I was interested in while drawing it.
I didn't draw it very clearly because I didn't feel a need to; but, as
such, I'm the only person who knows what I really meant with it. I'm
just not comfortable with keeping it there for anybody to misinterpret.